+++ package.json
@@ -69,6 +69,7 @@
     "@storybook/storybook-deployer": "^2.8.12",
     "@testing-library/react": "^12",
     "@types/jest": "^27",
+    "@types/lodash": "^4.14.184",
     "@types/node": "^18.7.14",
     "@types/react": "^16",
     "@types/react-dom": "^16",
